refactor: Apply SRP to ModuleExecutor and ModuleScheduler (#1421)
Extract single-responsibility components from large orchestration classes:

ModuleExecutor (779→218 lines, 21→9 dependencies):
- IModuleResultRegistrar: Registers module execution results
- IParallelLimitHandler: Manages parallel execution semaphore
- IDependencyWaiter: Waits for module dependencies to complete
- IModuleLifecycleEventInvoker: Invokes module lifecycle hooks
- IModuleRunner: Executes individual modules with retry/timeout
- IAlwaysRunHandler: Handles AlwaysRun modules after pipeline failure

ModuleScheduler (743 lines simplified):
- IModuleConstraintEvaluator: Evaluates NotInParallel and lock key constraints

Benefits:
- Improved testability through focused interfaces
- Better maintainability with smaller, single-purpose classes
- Enhanced extensibility via dependency injection
